BREAKING: rename templates directory to theme
This is to make room for templates that will be used by the new-post
command (#8).

This is a breaking change because previously we expected to find the
Tera templates in the templates directory. Instead, by default they are
not expected to be in the theme directory, but this location can now
be overridden by the `theme` key in the `Site.toml` file.